+++ amethyst_audio/src/components/audio_listener.rs
@@ -1,5 +1,5 @@
 use amethyst_core::{
-    nalgebra::Vector3,
+    nalgebra::Point3,
     specs::{prelude::Component, storage::HashMapStorage},
 };
 
@@ -11,9 +11,9 @@ pub struct AudioListener {
     /// Output used by this listener to emit sounds to
     pub output: Output,
     /// Position of the left ear relative to the global transform on this entity.
-    pub left_ear: Vector3<f32>,
+    pub left_ear: Point3<f32>,
     /// Position of the right ear relative to the global transform on this entity.
-    pub right_ear: Vector3<f32>,
+    pub right_ear: Point3<f32>,
 }
 
 impl Component for AudioListener {